% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_shp}
\alias{boston_shp}
\title{Shapefile of Boston grid cells}
\format{
tbd.
\describe{
\item{id}{Numeric, id for each gridcell}
\item{geometry}{Geometry, spatial geometry for each polygon grid cell}
}
}
\usage{
boston_shp
}
\description{
Polygons and ids that correspond to `boston_data` ids
}
\keyword{datasets}
