% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.cityModel.R
\name{head.cityModel}
\alias{head.cityModel}
\title{head.cityModel}
\usage{
\method{head}{cityModel}(x, ...)
}
\arguments{
\item{x}{a cityModel object}

\item{...}{other arguments passed to plot}
}
\description{
head.cityModel
}
