% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cityModel.R
\name{plot.cityModel}
\alias{plot.cityModel}
\title{plot.cityModel}
\usage{
\method{plot}{cityModel}(x, feature = NA, hod = NA, doy = NA, ...)
}
\arguments{
\item{x}{a cityModel object}

\item{feature}{a column in the cityModel object to plot}

\item{...}{other arguments passed to plot}
}
\description{
plot.cityModel
}
