% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cityModel.R
\name{print.cityModel}
\alias{print.cityModel}
\title{print.cityModel}
\usage{
\method{print}{cityModel}(x, ...)
}
\arguments{
\item{x}{a cityModel object}

\item{...}{other arguments passed to print}
}
\description{
print.cityModel
}
