\name{smithCEE2025-package}
\alias{smithCEE2025-package}
\alias{smithCEE2025}
\docType{package}

\title{
  Code from Smith 2025 (smithCEE2025)
}

\description{
  The package \pkg{smithCEE2025} contains code from smithCEE2025.
}

\section{Modelling framework}{
  Some text here
}

\section{Additional information}{

  The vignette \strong{boston_tree_cover} offers a general illustration.

  A vignette is available by typing:

    \code{vignette("boston_tree_cover")}

  Please use \code{citation("smithCEE2025")} to cite this package.
}

\author{
  Chad Milando

  Maintainer: Chad Milando <\email{cmilando@bu.edu}>
}

\references{
  Smith, Ian A., Dan Li, David K. Fork, Gregory A. Wellenius, and Lucy R. Hutyra. "Integrated tree canopy expansion and cool roofs can optimize air temperature and heat exposure reductions in Boston." Communications Earth & Environment 6, no. 1 (2025): 507.

}

\keyword{package}


