% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_data}
\alias{boston_data}
\title{Boston feature data}
\format{
tbd.
\describe{
\item{Cases}{Numeric}
}
}
\usage{
boston_data
}
\description{
For every gridcell for every time point, feature data
}
\keyword{datasets}
