% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_shp}
\alias{boston_shp}
\title{boston shapefile}
\format{
tbd.
\describe{
\item{Cases}{Numeric}
}
}
\usage{
boston_shp
}
\description{
For every gridcell for every time point, feature data
}
\keyword{datasets}
