% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_output.R
\name{retrieve_output}
\alias{retrieve_output}
\title{Retrieve output}
\usage{
retrieve_output(city_name = "boston")
}
\arguments{
\item{city_name}{the name of a city, converts to lowercase}

\item{clip}{optional, a shapefile to clip the output to}
}
\description{
Retrieve output
}
