% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station_data}
\alias{station_data}
\title{Weather station data}
\format{
tbd.
\describe{
\item{Cases}{Numeric}
}
}
\usage{
station_data
}
\description{
For every gridcell for every time point, feature data
}
\keyword{datasets}
