% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_city_names}
\alias{valid_city_names}
\title{Valid city names}
\format{
tbd.
\describe{
\item{Cases}{Numeric}
}
}
\usage{
valid_city_names
}
\description{
For now, just boston
}
\keyword{datasets}
